% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.block_from_list}
\alias{.block_from_list}
\title{Convert a raw API block list into a typed Block S3 object.}
\usage{
.block_from_list(d)
}
\description{
Convert a raw API block list into a typed Block S3 object.
}
\keyword{internal}
