% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{.parse_result_from_list}
\alias{.parse_result_from_list}
\title{Convert a raw API response into a ParseResult S3 object.}
\usage{
.parse_result_from_list(d)
}
\description{
Convert a raw API response into a ParseResult S3 object.
}
\keyword{internal}
